Extract the data device helper classes to their own implementation file
